pyproject.toml
[project]
name = "alert-collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "celery>=5.5.3",
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.11.0",
    "prometheus-client>=0.23.1",
    "sqlalchemy>=2.0.44",
    "structlog>=25.4.0",
    "typer>=0.19.2",
    "uvicorn>=0.38.0",
]

[project.scripts]
alert-collector = "alert_collector.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "respx>=0.23.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
